Modulo operator does not accept doubles as arguments gg. fixed by turning x into an int from double

+++ week02/arithmetic.cpp
@@ -20,7 +20,7 @@ using namespace std;
  * 
  */
 int main(int argc, char** argv) {
-    double x;
+    int x;
     int choice;
     cout << "Please input a number for x: ";
     cin >> x;
